third_party/blink/web_tests/virtual/parse-processing-instructions-as-bogus-comments/html5lib/generated/run-tests1-write-expected.txt

CONSOLE ERROR: Uncaught SyntaxError: Failed to execute 'write' on 'Document': Unexpected token '<'
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select>. Please add the missing </select> end tag.
CONSOLE WARNING: A <select> tag was parsed within another <select> tag and was converted into </select>. Please add the missing </select> end tag.
../resources/tests1.dat: PASS
